A workaround for dropped output
This is quite tricky problem, because the *correct* resolution would require us to add some form of race-free wait for the process to start before we try collecting its outputs. Hopefully we can rely on the rules of go routines preemption to ensure that both go routines will be preempted no sooner than Read call inside Scan, but I do not feel so certain about it.

Still, this should fix some of the transient test failures we are seeing.