Add WaitForResource, WaitForResources
These functions provide a convenient shortcut when we already have a resource the like of which we are waiting for, or if we want to perform some further filtering of the resources e.g. based on status.

Unit tests cannot not be implemented at this stage, this will come after we have brought some mock kubernetes into the project