Refactoring of WaitFor to use int rather than *int32
This is breaking change, but getting over if will actually simplify majority of uses of WaitFor. The intent of the original design was to allow the users to directly pass e.g. deployment.Spec.Replicas which is *int32, but this turned out not be be very useful, while at the same time creating friction for simpler case of just expecting a fixed number of pods to show up. The original design has also spawned the helper function WaitForNone which is no longer needed, because the user can just as well pass 0 to WaitFor.