Add support for default options and os.Args override
Users tests should use this to set default parameters before parsing. The OverrideOsArgs is meant to aid unit testing of Parse() function.

This change makes use of Functional Options pattern: Parse() takes a variadic slice of ParseOptionFn functions, each of which applies some modification to ParseOptions, and we provide a selection of higher-order functions which generate ParseOptionFn for modification of each field in ParseOptions.