Keep TEST_OPERATOR_NS and TEST_COUNT separate for each test
These two variables are specific for each test, so if we simply store them in the current os.Environ() with os.Setenv, then one test will overwrite another if they happen to run concurrently. We have to be smarter about it and set these variables in the forked make glue process instead.